+++ src/routes/shopRoutes.js
@@ -7,12 +7,12 @@ const shopControllers = new shopController();
 router
   .get("/", shopControllers.shopGet)
 
-  // .get("/cart", shopControllers.shopCartGet)
-  // .post("/cart", shopControllers.shopCartPost)
-  // .delete("/cart", shopControllers.shopCartDelete)
-
+  .get("/cart", shopControllers.shopCartGet)
+  .post("/cart", shopControllers.shopCartPost)
+  .delete("/cart", shopControllers.shopCartDelete)
+  
   .get("/item/:id", shopControllers.itemIdGet)
 
-  // .post("/item/:id/add", shopControllers.itemIdAddPost);
+  .post("/item/:id/add", shopControllers.itemIdAddPost);
 
 export default router;
