+++ src/views/admin/admin.ejs
@@ -27,7 +27,6 @@
                         class="search_input"
                         type="text"
                         placeholder="código, nombre o categoria"
-                        onkeyup="filterProducts(this.value)"
                     />
                 </label>
             </article>
